+++ labmanager-interface/src/app/components/edit/edit-examination/edit-examination.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { verifyGeneric, verifyText } from 'src/app/shared-methods/validations';
 import { BaseExaminationComponent } from '../../shared/base-examination/base-examination.component';
 
 @Component({
@@ -11,6 +12,8 @@ export class EditExaminationComponent extends BaseExaminationComponent implement
   examinationDetails: any;
   
   ngOnInit() {
+    this.localStorageLabId = localStorage.getItem('labId');
+
     this.route.params
     .subscribe((params) => {
       if(params && params.id) {
@@ -21,6 +24,9 @@ export class EditExaminationComponent extends BaseExaminationComponent implement
             this.examinationDetails = result;
 
             this.initForm();
+
+            this.getPatients();
+            this.getExaminationTypes();
           });
       }
     });
@@ -28,51 +34,48 @@ export class EditExaminationComponent extends BaseExaminationComponent implement
 
   initForm() {
     this.examinationForm = this._formBuilder.group({
-      patient_id: [this.examinationDetails.patient_id],
-      examinationType_id: [this.examinationDetails.examinationType_id],
+      id: [this.examinationId],
       information: [this.examinationDetails.information],
       completed: [this.examinationDetails.completed],
       madeOnDate: [this.examinationDetails.madeOnDate],
-      employee_id: [this.examinationDetails.employee_id],
-      laboratory_id: [this.examinationDetails.laboratory_id],
     });
   }
 
   save() {
     if (this.examinationForm.valid) {
-      const madeOnDate = this.convertDate();
-      const patient_id = this.examinationForm.value.patient_id;
-      const examinationType_id = this.examinationForm.value.examinationType_id;
+      const madeOnDate = this.convertDate(this.examinationForm.value.madeOnDate);
+      const id = this.examinationForm.value.id;
       const information = this.examinationForm.value.information;
       const completed = this.examinationForm.value.completed;
-      const employee_id = this.examinationForm.value.employee_id;
-      const laboratory_id = this.examinationForm.value.laboratory_id;
 
-      const examination = {
-        madeOnDate: madeOnDate,
-        patient_id: patient_id,
-        information: information,
-        completed: completed,
-        employee_id: employee_id,
-        examinationType_id: examinationType_id,
-        laboratory_id: laboratory_id,
-      };
-
-      this.examinationService.updateExamination(examination)
-        .subscribe((result: any)=> {
-          if (result) {
-            window.location.href = '/Get/Results';
-          }
-        });
+      if (verifyGeneric(madeOnDate) && verifyGeneric(id) && verifyGeneric(completed) && verifyText(information)) {
+        const examination = {
+          madeOnDate: madeOnDate,
+          id: id,
+          information: information,
+          completed: completed,
+        };
+  
+        this.examinationService.updateExamination(examination)
+          .subscribe((result: any)=> {
+            if (result) {
+              window.location.href = '/Get/Results/All';
+            }
+          });
+      } else {
+        console.log('Data failed verification!');
+      }
     } else {
       console.log('Invalid Form!');
     }
   }
 
   remove() {
-    this.examinationService.removeExamination(this.examinationId)
-      .subscribe((result: any)=> {
-        window.location.href = '/Get/Results';
-    });
+    if (verifyGeneric(this.examinationId)) {
+      this.examinationService.removeExamination(this.examinationId)
+        .subscribe((result: any)=> {
+          window.location.href = '/Get/Results';
+        });
+    }
   }
 }
